requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
torch==2.2.1
torchvision==0.17.1
opencv-python==4.9.0.80
numpy==1.26.3
Werkzeug==3.0.1
rembg==2.0.50
Pillow==10.1.0
python-socketio==5.11.1